chore: move html logic into its own file
